# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pikhlak API
  description: |
    Headless API для Telegram AI-бота автосервиса Pikhlak.

    - **Admin API** — JWT Bearer для Vue Admin Panel
    - **Bot API** — HMAC подпись для lightweight Telegram service
    - **Webhooks** — Telegram → Laravel (ingress)
  version: 1.0.0
  contact:
    name: Pikhlak Platform
servers:
  - url: https://api.pikhlak.local/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /admin/cars/{carId}:
    get:
      tags: [Cars]
      parameters:
        - $ref: '#/components/parameters/CarId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
    patch:
      tags: [Cars]
      parameters:
        - $ref: '#/components/parameters/CarId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
    delete:
      tags: [Cars]
      parameters:
        - $ref: '#/components/parameters/CarId'
      responses:
        '204':
          description: Car archived
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin SPA access token
    BotHmacAuth:
      type: apiKey
      in: header
      name: X-Bot-Signature
      description: HMAC-SHA256 with bot API secret
  parameters:
...
    CarId:
      name: carId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        last_page:
          type: integer
...
    Car:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
          format: uuid
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        price:
          type: number
        currency:
          type: string
        status:
          type: string
        specs:
          type: object
          additionalProperties: true
...
    CarResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Car'
...
    CarUpdateRequest:
      type: object
      properties:
        status:
          type: string
        price:
          type: number
        specs:
          type: object
...